Introduce concept of signer BREAKING CHANGE
This changes the key pair to be a signer.

A signer is in the default case the same as the previous key pair map
either bound or in an atom.

However using the multi method `(sign-with-signer)` we can now
implement alternative signing strategies such as metamask, ethereum
urls or uport.

An optional `:type` attribute in the `signer` is used to select the
method used for signing.

changes `(keypair)` to `(current-signer)`, `(bound-keypair)` to
`(bound-signer)` and `global-keypair` to `global-signer`.